Added the core components as defined in design.
The challenge here is to make a model that behaves properly over a time
frame, as this is a simulation. Multithreading has been considered
during this phase, but it's implementation will take longer.

This commit includes all the components that will act as "business
model", and some utils to help them behave as expected.