Major changes needed.
The current model didn't consider some logic required for networking to
simulate some conditions in a proper way. So, deleting everything again
and starting over.

This time, some BDD techniques will be added to the project, so it will
be easier to change the model in the future (if needed).