Simulation feature has been added to the tree.
Added some of the classes that will lay the foundation for the
simulation feature, but some refactoring might be requred in other
modules.

These changes may result in redundancy with some classes in the "networking" module, so a refactor will be required. Such as the Network class that is supposed to hold references to all the nodes in the simulation, which can actually be handled by the simulation module. That way the network graph would be built without specifically asking for it.