Removed the Network class and denpendants
Network is going to be in the simulation module, and it si going to be
called Plane, as the actual network will be built implicitly.

The Plane object will be responsible for holding references to every
node in the simulation. Additionally, it will be responsible for
managing every aspect of the nodes, such as threading.